+++ apisix/admin/init.lua
@@ -455,7 +455,7 @@ function _M.init_worker()
     router = route.new(uri_route)
 
     -- register reload plugin handler
-    events = require("apisix.event")
+    events = require("apisix.events")
     events:register(reload_plugins, reload_event, "PUT")
 
     if ngx_worker_id() == 0 then
